feat: VoicesRichText - extract resizable_box_parent.dart
