+++ catalyst_voices/lib/widgets/text_field/voices_text_field.dart
@@ -1,5 +1,6 @@
 import 'dart:math';
 
+import 'package:catalyst_voices/widgets/common/resizable_box_parent.dart';
 import 'package:catalyst_voices_assets/catalyst_voices_assets.dart';
 import 'package:catalyst_voices_brands/catalyst_voices_brands.dart';
 import 'package:catalyst_voices_shared/catalyst_voices_shared.dart';
@@ -156,7 +157,7 @@ class _VoicesTextFieldState extends State<VoicesTextField> {
           ),
           const SizedBox(height: 4),
         ],
-        _ResizableBoxParent(
+        ResizableBoxParent(
           resizable: resizable,
           child: TextFormField(
             textAlignVertical: TextAlignVertical.top,
@@ -569,98 +570,4 @@ class VoicesTextFieldDecoration {
     this.counterText,
     this.showStatusSuffixIcon = true,
   });
-}
-
-class _ResizableBoxParent extends StatelessWidget {
-  final bool resizable;
-  final Widget child;
-
-  const _ResizableBoxParent({
-    required this.resizable,
-    required this.child,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    if (!resizable) {
-      return child;
-    }
-
-    return LayoutBuilder(
-      builder: (BuildContext context, BoxConstraints constraints) {
-        return _ResizableBox(
-          constraints: constraints,
-          child: child,
-        );
-      },
-    );
-  }
-}
-
-class _ResizableBox extends StatefulWidget {
-  final BoxConstraints constraints;
-  final Widget child;
-
-  const _ResizableBox({
-    required this.constraints,
-    required this.child,
-  });
-
-  @override
-  State<_ResizableBox> createState() => _ResizableBoxState();
-}
-
-class _ResizableBoxState extends State<_ResizableBox> {
-  static const double _minWidth = 40;
-  static const double _minHeight = 40;
-
-  late double _width;
-  late double _height;
-
-  @override
-  void initState() {
-    super.initState();
-
-    _width = widget.constraints.maxWidth != double.infinity
-        ? widget.constraints.maxWidth
-        : widget.constraints.constrainWidth(_minWidth);
-
-    _height = max(widget.constraints.minHeight, _minHeight);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      children: [
-        SizedBox(
-          width: _width,
-          height: _height,
-          child: widget.child,
-        ),
-        Positioned(
-          bottom: 0,
-          right: 0,
-          child: MouseRegion(
-            cursor: SystemMouseCursors.resizeDownRight,
-            child: GestureDetector(
-              onPanUpdate: (details) {
-                setState(() {
-                  _width = (_width + details.delta.dx).clamp(
-                    _minWidth,
-                    widget.constraints.maxWidth,
-                  );
-
-                  _height = (_height + details.delta.dy).clamp(
-                    _minHeight,
-                    widget.constraints.maxHeight,
-                  );
-                });
-              },
-              child: VoicesAssets.images.dragger.buildIcon(size: 15),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
 }
